fix: handle lobby_members before deleting lobbies

+++ frontend/src/services/deleteGroup.ts
@@ -1,6 +1,26 @@
 import { supabase } from "../shared/api/supabaseClient";
 
 export async function deleteGroup(groupId: string) {
+    const { data: lobbies } = await supabase
+        .from("lobbies")
+        .select("id")
+        .eq("group_id", groupId);
+
+    const lobbyId = lobbies?.map(lobby => lobby.id) ?? [];
+    const { error: deleteLobbiesMembersError } = await supabase
+        .from('lobbies_members')
+        .delete()
+        .eq('lobby_id', lobbyId);
+
+    if (deleteLobbiesMembersError) throw deleteLobbiesMembersError;
+
+    const { error: deleteLobbiesError } = await supabase
+        .from('lobbies')
+        .delete()
+        .eq('group_id', groupId);
+
+    if (deleteLobbiesError) throw deleteLobbiesError;
+
     const { error: deleteGroupMembersError } = await supabase
         .from('group_members')
         .delete()
